Refactor: extract complex cody/manage components
