+++ client/web/src/cody/onboarding/EditorStep.tsx
@@ -6,7 +6,7 @@ import type { TelemetryRecorder } from '@sourcegraph/shared/src/telemetry'
 import { H2, Text, H5 } from '@sourcegraph/wildcard'
 
 import type { IEditor } from './CodyOnboarding'
-import { editorGroups } from './editorGroups'
+import { editorGroups } from '../editorGroups'
 
 import styles from './CodyOnboarding.module.scss'
